requirements.txt
pymupdf
python-dotenv
openai
requests
psycopg2-binary
fastapi
uvicorn
psycopg2
jinja2